sort choices alphabetically
Choices were displayed in the order they appeared in the schema.
This was fine when choices were displayed in a single language
but didn't work so well with multiple languages.

With this update, setting sorted_choices:true in the schema will cause
choices to be displayed in alphabetical order, no matter the language in
which it is displayed.