Use the choices list to determine size of select field
The previous code only works if you define a list of values in your
schema JSON. For the new choices_helper options, where field.choices is
not set, the size would always be set to 0.

This change uses the correct choices list to determine the correct size
of the select field.