Enable Travis
The extension is currently being tested in Circle CI, which wwe'd have
to upgrade to v2.0 soon. Moreover it's just testing CKAN 2.4 to 2.6,
and as this is a pretty critical extension we should test across
all CKAN versions. I'm sure there are ways to do this in Circle CI
but we have a really well established pattern using Travis so it makes
sense to use it for now.