Pass dataset name to resource fields snippets
Resource field snippets in the new form have no way of knowing the
dataset id/name (if it's an edit form there is `data.package_id`).

These changes are not the most elegant, because 1) they add a
resource-specific key to the generic `form_field.html` and 2) they pass
the dataset name instead of the id.

This is because, annoyingly, upstream CKAN core only passes this from
`new_resource.html` / `new_resource_not_draft.html` to
`resource_form.html`, and not `pkg_dict`