Provide a new plugin for indexing repeating subfields
The `scheming_subfields_index` plugin will group the values of the
same subfields in a text field that will make the values findable.
They are indexed as `extras_{field_name}__{key}`. `extras_*` is a
dynamic `text` Solr field that will allow free-text search on these
values.

Added tests and updated the docs.